pyproject.toml
[project]
name = "lttng-tools-benchmarks"
version = "0.1.0"
description = "Blackbox benchmarks for LTTng-tools"
authors = [
    {name = "Kienan Stewart",email = "kstewart@efficios.com"}
]
license = {text = "GPL-2.0-only"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml (>=6.0.3,<7.0.0)"
]

[project.scripts]
cli = "lttng-tools-benchmarks.benchmark"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]

[tool.isort]
profile = "black"
